fix fanctionality
