--- client/src/components/massages/Navbar.js
@@ -1,23 +1,0 @@
-// import React from "react";
-// import "./navbar.css";
-
-// export default function Navbar() {
-// 	return (
-// 		<nav className="navbar">
-// 			<ul className="menu ">
-// 				<li className="menu-item">
-// 					<a href="#">Friends</a>
-// 				</li>
-// 				<li className="menu-item">
-// 					<a href="#">Profile</a>
-// 				</li>
-// 				<li className="menu-item">
-// 					<a href="#">Message</a>
-// 				</li>
-// 				<li className="menu-item">
-// 					<a href="#">Corrections</a>
-// 				</li>
-// 			</ul>
-// 		</nav>
-// 	);
-// }
