--- client/src/components/massages/manageNav/Nav.js
@@ -1,5 +1,0 @@
-import React from "react";
-
-export default function Nav() {
-	return <div>Nav</div>;
-}
