+++ client/src/components/massages/massage.css
@@ -22,6 +22,7 @@
 
 .headerMessage {
 	width: 100%;
+	background-color: #fff;
 	font-size: 18px;
 	border: 1px solid rgb(196, 206, 215);
 }
