+++ package.json
@@ -66,18 +66,18 @@
 		"react": "^18.2.0",
 		"react-bootstrap": "2.7.2",
 		"react-dom": "^18.2.0",
+		"react-icons": "^4.8.0",
 		"react-router": "6.8.2",
 		"react-router-dom": "^6.9.0",
 		"react-select": "^5.7.0",
+		"react-toastify": "^5.5.0",
 		"sass": "^1.58.0",
 		"style-loader": "^3.3.1",
 		"wait-on": "^7.0.1",
 		"webpack": "^5.75.0",
 		"webpack-cli": "^5.0.1",
 		"webpack-dev-server": "^4.11.1",
-		"webpack-merge": "^5.8.0",
-		"react-icons": "^4.8.0",
-		"react-toastify": "^5.5.0"
+		"webpack-merge": "^5.8.0"
 	},
 	"dependencies": {
 		"@babel/runtime": "^7.20.13",
